sales

+++ api/src/services/purchaseRequestServices.ts
@@ -61,6 +61,7 @@ export const createPurchaseRequest = async (data: {
           data: data.purchaseRequestItem.map((item) => ({
             material: item.material,
             quantity: item.quantity,
+            originalQuantity: item.quantity,
             unit: item.unit,
             requestedDate: new Date(item.requestedDate),
             description: item.description,
@@ -97,7 +98,13 @@ export const getPurchaseRequest = async (id: number) => {
       id,
     },
     include: {
-      personnel: true,
+      personnel: {
+        select: {
+          firstName: true,
+          lastName: true,
+          department: true,
+        },
+      },
       purchaseRequestItem: true,
     },
   });
@@ -106,7 +113,13 @@ export const getPurchaseRequest = async (id: number) => {
 export const getAllPurchaseRequests = async () => {
   return await prisma.purchaseRequest.findMany({
     include: {
-      personnel: true,
+      personnel: {
+        select: {
+          firstName: true,
+          lastName: true,
+          department: true,
+        },
+      },
     },
   });
 };
@@ -148,11 +161,13 @@ export const updatePurchaseRequest = async (
         upsert: data.purchaseRequestItem.map((item) => ({
           where: {
             id: item.id,
+            purchaseRequestId: id,
           },
 
           update: {
             material: item.material,
             quantity: item.quantity,
+            originalQuantity: item.quantity,
             unit: item.unit,
             requestedDate: new Date(item.requestedDate),
             description: item.description,
@@ -160,6 +175,7 @@ export const updatePurchaseRequest = async (
           create: {
             material: item.material,
             quantity: item.quantity,
+            originalQuantity: item.quantity,
             unit: item.unit,
             requestedDate: new Date(item.requestedDate),
             description: item.description,
@@ -171,12 +187,24 @@ export const updatePurchaseRequest = async (
 };
 
 export const deletePurchaseRequest = async (id: number) => {
+  if (
+    await prisma.purchaseOrder.findFirst({
+      where: {
+        purchaseRequestId: id,
+      },
+    })
+  ) {
+    return new Error("Purchase request has purchase orders, cannot delete");
+  }
+  await prisma.purchaseRequestItem.deleteMany({
+    where: {
+      purchaseRequestId: id,
+    },
+  });
+
   return await prisma.purchaseRequest.delete({
     where: {
       id,
-    },
-    include: {
-      purchaseRequestItem: true,
     },
   });
 };
@@ -190,10 +218,15 @@ export const getSubordinatesPurchaseRequests = async (personnelId: number) => {
       personnel: {
         supervisorId: personnelId,
       },
-      approvalFromPurchasing: null,
     },
     include: {
-      personnel: true,
+      personnel: {
+        select: {
+          firstName: true,
+          lastName: true,
+          department: true,
+        },
+      },
     },
   });
 };
@@ -220,6 +253,21 @@ export const supervisorApproval = async (id: number, approval: boolean) => {
     },
   });
   if (res) {
+    const items = await prisma.purchaseRequestItem.findMany({
+      where: {
+        purchaseRequestId: id,
+      },
+    });
+    const updateItems = items.map((item) => {
+      return prisma.purchaseRequestItem.update({
+        where: {
+          id: item.id,
+        },
+        data: {
+          supervisorQuantity: approval ? item.quantity : 0,
+        },
+      });
+    });
     supervisorApprovalNotification(id, approval);
   }
   return res;
@@ -286,10 +334,15 @@ export const getSupervisorApprovedPurchaseRequests = async () => {
   return await prisma.purchaseRequest.findMany({
     where: {
       approvalFromSupervisor: true,
-      approvalFromManagement: null,
     },
     include: {
-      personnel: true,
+      personnel: {
+        select: {
+          firstName: true,
+          lastName: true,
+          department: true,
+        },
+      },
     },
   });
 };
@@ -310,11 +363,6 @@ const purchasingApprovalNotification = async (
       },
     },
   });
-  const management = await prisma.personnel.findMany({
-    where: {
-      department: "yon",
-    },
-  });
   const personnel = purchaseRequest?.personnel;
   const supervisor = personnel?.supervisor;
   if (personnel && supervisor) {
@@ -328,20 +376,6 @@ const purchasingApprovalNotification = async (
         category: "Personnel",
         link: "/purchase-request/my-requests/view/?id=" + id,
       },
-    });
-  }
-
-  if (management && approval) {
-    const buyingNot = await prisma.notification.createMany({
-      data: management.map((man) => ({
-        personnelId: man.id,
-        title: `Yeni bir satın alma talebi satın alma departmanı tarafından onaylandı.`,
-        description: `Satın alma departmanı bir satın alma talebi ${
-          approval ? "onayladı, onayınızı bekliyor" : "reddetti"
-        }.`,
-        category: "Personnel",
-        link: "/purchase-request/management/view/?id=" + id,
-      })),
     });
   }
 };
@@ -358,101 +392,24 @@ export const purchasingApproval = async (id: number, approval: boolean) => {
     },
   });
   if (req) {
+    const items = await prisma.purchaseRequestItem.findMany({
+      where: {
+        purchaseRequestId: id,
+      },
+    });
+    const updateItems = items.map((item) => {
+      return prisma.purchaseRequestItem.update({
+        where: {
+          id: item.id,
+        },
+        data: {
+          purchasingQuantity: approval ? item.quantity : 0,
+        },
+      });
+    });
     purchasingApprovalNotification(id, approval);
   }
   return req;
-};
-
-// Stage three (management approval)
-export const getPurchaseRequestsForManagement = async () => {
-  return await prisma.purchaseRequest.findMany({
-    where: {
-      approvalFromSupervisor: true,
-      approvalFromPurchasing: true,
-    },
-    include: {
-      personnel: true,
-    },
-  });
-};
-
-export const managementApproval = async (
-  id: number,
-  data: {
-    approvalFromManagement: boolean;
-  }
-) => {
-  const req = await prisma.purchaseRequest.update({
-    where: {
-      id,
-      approvalFromSupervisor: true,
-      approvalFromPurchasing: true,
-    },
-    data: {
-      approvalFromManagement: data.approvalFromManagement,
-      approvalFromManagementDate: new Date(),
-    },
-  });
-  if (req) {
-    managementApprovalNotification(id, data.approvalFromManagement);
-  }
-  return req;
-};
-
-const managementApprovalNotification = async (
-  id: number,
-  approval: boolean
-) => {
-  const purchaseRequest = await prisma.purchaseRequest.findUnique({
-    where: {
-      id,
-    },
-    include: {
-      personnel: {
-        include: {
-          supervisor: true,
-        },
-      },
-    },
-  });
-  const buying = await prisma.personnel.findMany({
-    where: {
-      department: "stn",
-    },
-  });
-  const personnel = purchaseRequest?.personnel;
-  const supervisor = personnel?.supervisor;
-  if (personnel && supervisor) {
-    const personnelNot = await prisma.notification.create({
-      data: {
-        personnel: { connect: { id: personnel.id } },
-        title: `Üst yönetim satın alma talebinizi ${
-          approval ? "onayladı" : "reddetti"
-        }.`,
-        description: `Satın alma talebiniz üst yönetim tarafından ${
-          approval
-            ? "onaylandı ve satın alma departmanı tarafından satın alınacak"
-            : "reddedildi"
-        }.`,
-        category: "Personnel",
-        link: `/purchase-request/my-requests/view/?id=${id}`,
-      },
-    });
-  }
-
-  if (buying && approval) {
-    const buyingNot = await prisma.notification.createMany({
-      data: buying.map((buy) => ({
-        personnelId: buy.id,
-        title: `Yeni bir satın alma talebi Üst Yönetim tarafından onaylandı.`,
-        description: `Üst Yönetim bir satın alma talebi ${
-          approval ? "onayladı, sipariş oluşturmanızı bekliyor" : "reddetti"
-        }.`,
-        category: "Personnel",
-        link: "/purchase-request/view/?id=" + id,
-      })),
-    });
-  }
 };
 
 // private views
@@ -462,7 +419,13 @@ export const getPurchaseRequestsByPersonnel = async (personnelId: number) => {
       personnelId,
     },
     include: {
-      personnel: true,
+      personnel: {
+        select: {
+          firstName: true,
+          lastName: true,
+          department: true,
+        },
+      },
     },
   });
 };
